[PORT FROM MAIN] Fix the timestamp overflow bug
BZ: 28089 50102

This patch fix an Android defeat. 32 bits timestamp is not big enough,
change to 64 bits to avoid overflow.

Original patch:
R3: 41774

Orig-Change-Id: I17d8d84cdec84167d4b80ff2a55ed35be2419d53
Category: aosp improvement
Domain: Video.Media-local playback
Origin:
Upstream-Candidate: yes

+++ media/libstagefright/MPEG4Extractor.cpp
@@ -443,7 +443,7 @@ sp<MetaData> MPEG4Extractor::getTrackMetaData(
                 }
             } else {
                 uint32_t sampleIndex;
-                uint32_t sampleTime;
+                int64_t sampleTime;
                 if (track->sampleTable->findThumbnailSample(&sampleIndex) == OK
                         && track->sampleTable->getMetaDataForSample(
                             sampleIndex, NULL /* offset */, NULL /* size */,
@@ -3085,7 +3085,7 @@ status_t MPEG4Source::read(
                     sampleIndex, &syncSampleIndex, findFlags);
         }
 
-        uint32_t sampleTime;
+        int64_t sampleTime;
         if (err == OK) {
             err = mSampleTable->getMetaDataForSample(
                     sampleIndex, NULL, NULL, &sampleTime);
@@ -3131,7 +3131,7 @@ status_t MPEG4Source::read(
 
     off64_t offset;
     size_t size;
-    uint32_t cts;
+    int64_t cts;
     bool isSyncSample;
     bool newBuffer = false;
     if (mBuffer == NULL) {
